+++ mlp/mlp_model.py
@@ -13,4 +13,4 @@ class MlpModel(BaseModel):
             
             predicted_indicies = torch.argmax(logits, dim=1)
             predicted_actions = (predicted_indicies - 1).cpu().numpy().squeeze()
-            return predicted_actions
+            return predicted_actions.astype(np.int8)
